# Optimize xgettext performance