This should not result in any external behavioral difference. The goal is just to speed up the  `fish_xgettext.fish` script, and to clean it up a bit.

It takes my machine currently about 7 seconds to run the version of the script on master, and about 5.5 seconds on this branch. (Assuming cargo does not have to compile anything.) This is an absolute speedup of 1.5 seconds, or a relative one of 1.27.

If excluding the time it takes to run cargo-expand (about 4.5 seconds) it takes the rest of the script from 2.5 seconds to 1 second (relative speedup 2.5).

The `messages.pot` file produced by the script is unchanged.